Add LogLink docs & example.
