+++ doc/en/download/Assertions.rst
@@ -9,11 +9,60 @@ Basic
 .. _example_assertions_basic:
 
 Required files:
-  - :download:`test_plan.py <../../../examples/Assertions/Basic/test_plan.py>`
+  - :download:`test_plan_basic.py <../../../examples/Assertions/Basic/test_plan_basic.py>`
 
-test_plan.py
-++++++++++++
-.. literalinclude:: ../../../examples/Assertions/Basic/test_plan.py
+test_plan_basic.py
+++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_basic.py
+
+Required files:
+  - :download:`test_plan_group.py <../../../examples/Assertions/Basic/test_plan_group.py>`
+
+test_plan_group.py
+++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_group.py
+
+Required files:
+  - :download:`test_plan_exception.py <../../../examples/Assertions/Basic/test_plan_exception.py>`
+
+test_plan_exception.py
+++++++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_exception.py
+
+Required files:
+  - :download:`test_plan_dict.py <../../../examples/Assertions/Basic/test_plan_dict.py>`
+
+test_plan_dict.py
++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_dict.py
+
+Required files:
+  - :download:`test_plan_fix.py <../../../examples/Assertions/Basic/test_plan_fix.py>`
+
+test_plan_fix.py
+++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_fix.py
+
+Required files:
+  - :download:`test_plan_regex.py <../../../examples/Assertions/Basic/test_plan_regex.py>`
+
+test_plan_regex.py
+++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_regex.py
+
+Required files:
+  - :download:`test_plan_table.py <../../../examples/Assertions/Basic/test_plan_table.py>`
+
+test_plan_table.py
+++++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_table.py
+
+Required files:
+  - :download:`test_plan_xml.py <../../../examples/Assertions/Basic/test_plan_xml.py>`
+
+test_plan_xml.py
+++++++++++++++++
+.. literalinclude:: ../../../examples/Assertions/Basic/test_plan_xml.py
 
 
 Summarization
